Add development environment configuration, imported from kayobe
Use default value for controller_bootstrap_user

The default is to use the same username as the operator for
bootstrapping the controller. In vagrant this will still typically be
correct - the vagrant user. This should also be correct in the OpenStack
CI environment - the zuul user will be used.

Only use eth1 interface on the controller if it exists

For testing we can use a bridge without any ports.

Add support for a seed host

The host uses the devicemapper docker storage driver, and the all-in-one
network interface is breth1, as for the controller.

Story: 2001655
Task: 6683

Enable virtualised compute in overcloud

Adds the all-in-one overcloud host controller0 to the [compute] group.

This will cause nova compute and related services to be deployed on this
host, allowing VMs to be created.

Use QEMU for virtualised compute

Most development environments will use nested virtualisation, and we
can't guarantee that nested KVM support is available. Use QEMU as a
lowest common denominator.

Use a virtual environment for remote operations

Use overlay Docker storage for seed CI

Bifrost is now able to support the overlay Docker storage driver.
Switch to using it instead of devicemapper with a logical volume mounted
on a loopback device.

Use only a single worker process

Reduce the control plane's memory footprint by limiting the number of
OpenStack worker processes to one per-service.

Remove Zuul configuration

Need to add to openstack-infra/project-config without this present.

Add Zuul configuration, fix Gerrit repo

Adds a pep8 job, and seed and overcloud deployment jobs from kayobe.

Also fixes Gerrit config, a yamllint warning and uses python3 for pep8.

Change-Id: Ie37214fb73533b4d042cf624727762d94908f55f
Depends-On: Ie8a3dc3dfba1ce0bd78ac676e3c4553461f45a05

Use virtualenv_path to set ansible_python_interpreter

Prior to Ansible 2.5 Jinja expansion was not supported in this variable.
Now that kayobe depends on 2.5, we can use the correct path.

Change-Id: I85f62ae80a8751a356cee5a56bc01493ae04d6f4
Story: 2002010
Task: 19648

Synchronise with kayobe-config for Queens

Synchronise configuration format with kayobe-config @
371937fa9053d77a4bbca88d830d60dc76628c62.

Change-Id: Ife6e3f251120edcf5f94ca677a51bd5fe21fdc6f

Add a neutron allocation pool

This prevents neutron from allocating addresses that overlap
will the other address pools.

Change-Id: I69f8ca98ca17ea09d100a9745349391f0de59705

Use TinyIPA deployment images

The coreos images do not appear to be functional.
I didn't spend too much time debugging why, but it
seemed like they hung on boot (but that may just be a
lack of console logging).

The TinyIPA images can deploy a server without issue.

TrivialFix

Change-Id: I97bc80e516e684fd24ad453d24c5d77316539fb0